social_media_crawler/requirements.txt
app-store-reviews-reader==1.2
atlassian-python-api==3.18.0
beautifulsoup4==4.9.3
blis==0.7.5
boto==2.49.0
bs4==0.0.1
dateparser
fastapi==0.75.2
feedparser==6.0.8
gnews==0.2.7
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-play-scraper==1.0.2
googleapis-common-protos==1.54.0
GoogleNews==1.6.6
ipykernel==6.19.4
loguru==0.6.0
mmh3==3.0.0
mypy==1.0.1
numpy>=1.21.4
openpyxl
pandas==1.3.5
pre-commit==3.0.4
psycopg2-binary==2.9.5
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
reddit-rss-reader
regex==2021.11.10
requests
rsa==4.8
spacy==3.2.1
sqlalchemy==1.4.46
sqlalchemy-databricks==0.2.0
tqdm==4.64.1
trafilatura
uvicorn[standard]==0.20.0
yake==0.4.8
yt-dlp==2023.2.17
